+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Support autoformly.validation.messages in SimpleSchema
+
 ## [0.2.1] - 2015-11-18
 ### Changed
 - Refactor all files to use Strict Dependency Injection
@@ -20,6 +24,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.1 - 2015-11-14
 
+[Unreleased]: https://github.com/wieldo/meteor-autoformly/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/wieldo/meteor-autoformly/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/wieldo/meteor-autoformly/compare/v0.1.0...v0.2.0 
 [0.1.0]: https://github.com/wieldo/meteor-autoformly/compare/v0.0.1...v0.1.0